audacity : upgrade and rewrite
This upgrades port:audacity to the current stable release 2.4.2 and
port:audacity-devel to 2.4.2.790 (aka 3.0.0 beta). The main difference
between these versions is that the latter introduces and uses a new,
sqlite3-based file format (.aup3) which the older version cannot read
(and the newer version cannot export to the older .aup format).

A new subport is introduced, audacity-manual, which installs a local
copy of the online manual inside the app bundle where the application
will find it.

Several behind-the-scenes changes:
- the build is now handled through cmake and ninja
- port:audacity can depend on wxWidgets-3.0 and wxWidgets-3.2 via variants,
  port:audacity-devel will only build against wxWidgets-3.2 (by choice).
- the preferred way to build and run Audacity is against a patched version
  of wxWidgets 3.1.3 which contains fixes not included upstreams plus some
  new functionality (including accessibility support on Mac). The cmake
  build system has logic to grab the patched sources or use a locally
  provided copy, build them, and embed the required libraries in the app
  bundle. This has little effect on the size (more on the build time...)
  Therefore, the default build of the ports now uses this solution to
  provide the best possible user experience. Users who insist on runnnig
  against a "system" wxWidgets can set the +wxsystem variant, in which
  case the choice(s) outlined above become available.

Fixes: #5693